Decompile func0f153c20

+++ src/include/game/game_1531a0.h
@@ -19,7 +19,7 @@ Gfx *func0f153a34(Gfx *gdl, u32 arg1, u32 arg2, u32 arg3, u32 arg4, u32 arg5);
 u32 func0f153ab0(void);
 u32 func0f153b40(void);
 u32 func0f153b6c(void);
-void func0f153c20(s32 x, s32 y, f32 arg2, s32 arg3);
+void func0f153c20(s32 x, s32 y, f32 arg2, u8 arg3);
 void func0f153c50(void);
 void func0f153c88(void);
 void func0f153ce8(s32 x1, s32 x2, u32 arg2);
